requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.3.1
djangorestframework==3.17.1
psycopg2-binary==2.9.12
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
uuid6==2025.0.1
